% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridoR-package.R
\docType{package}
\name{corridoR-package}
\alias{corridoR}
\alias{corridoR-package}
\title{corridoR: migration corridors and genetic range shifts}
\description{
Tools for modeling genetic connectivity along migration corridors and
forecasting how it shifts under climate change. Start with
\code{vignette("corridoR")} for a short version of the workflow; the full
six-step tutorial is at \url{https://paulmaier.github.io/corridoR/articles/tutorial.html}.
}
\section{Citation}{

If you use corridoR, please cite both the 2022 paper that introduced the
method (Maier et al. 2022, Heredity 129:257-272,
\doi{10.1038/s41437-022-00561-x}) and the software paper. Run
\code{citation("corridoR")} for both.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paulmaier/corridoR}
  \item \url{https://paulmaier.github.io/corridoR/}
  \item Report bugs at \url{https://github.com/paulmaier/corridoR/issues}
}

}
\author{
\strong{Maintainer}: Paul A. Maier \email{maierpa@gmail.com} (\href{https://orcid.org/0000-0003-0851-8827}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Paul A. Maier \email{maierpa@gmail.com} (\href{https://orcid.org/0000-0003-0851-8827}{ORCID}) [copyright holder]
}

}
\keyword{internal}
