% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridoR.R
\name{corridor_country}
\alias{corridor_country}
\title{corridor_country}
\usage{
corridor_country(country)
}
\arguments{
\item{country}{The name of the country}
}
\value{
Country's ISO code.
}
\description{
This function allows you to find and search the right country code associated with the Northern Corridor Data.
If no argument is filed, all countries' ISO code will be displayed.
}
\examples{
mycountry <- corridor_country()
mycountry <- corridor_country(country = "Canada")
mycountry <- corridor_country("Canada")

}
