% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridoR.R
\name{corridor_data}
\alias{corridor_data}
\title{corridoR}
\usage{
corridor_data(country = country_code, port = port_name)
}
\arguments{
\item{country}{Countries' ISO code.}

\item{port}{Port's name}
}
\value{
Data for the country and port requested
}
\description{
This function allows you to find and display the Northern Corridor data according to the selected parameters.
The results come from a study conducted on the Nüance-R platform.
If no arguments are filled, all data will be displayed.
}
\examples{
df <- corridor_data(country = "USA",port = "")
df <- corridor_data(country = "",port = "")
df <- corridor_data(country = "",port = "QUEBEC")





}
