% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corridoR.R
\name{corridor_port}
\alias{corridor_port}
\title{corridor_port}
\usage{
corridor_port(port)
}
\arguments{
\item{port}{The name of the port}
}
\value{
Port's name included in the Northern Corridor data
}
\description{
This function allows you to find and search port's name included in the Northern Corridor data. 
If no argument is filled, all ports included in the Northern Corridor data will be displayed.
}
\examples{
myport <- corridor_port()
myport <- corridor_port(port = "HOUSTON")
myport <- corridor_port("HOUSTON")

}
